Eliminate stage name from bucket configurations
The s3 event subscription in the serverless lambda function definition
on a bucket name including a stage name results in the creation of a
CloudFormation bucket definition name that includes the stage name in
the CloudFormation definition.  This prevents multiple stages from being
deployed.

This change eliminates the serverless s3 event subscriptions and replaces
them with raw CloudFormation ones.  This allows the S3 bucket definitions
in CloudFormation to use any name, and one that doesn't contain the
stage name.