Use OpenAPI typescript-fetch generator
We were previously using the typescript-node generator which generated
code using the obsolete `request` API. This generator uses the modern
`fetch` API.

This change also requires us to enable the DOM library which makes a
number of fetch related types visible. In our case we directly depend on
the `RequestCredentials` type. Unfortunately we have to explicitly list
all libraries that we want to enable, we can't just add `DOM` to the
existing list. So we have to copy the `lib `entry from the imported
`tsconfig.json` and manually add the `DOM` library.